@@ -1,37 +1,37 @@
 ﻿using BenchmarkDotNet.Attributes;
 
-namespace PerfTesting.Trivia
+namespace PerfTesting.Trivia;
+
+[MemoryDiagnoser, ShortRunJob]
+public class MathTrivia
 {
-    public class MathTrivia
+    private static readonly IReadOnlyList<int> IntCollection = Enumerable.Range(0, 100).ToList().AsReadOnly();
+
+    [Benchmark]
+    public bool ModuloIsEven()
     {
-        private static readonly IReadOnlyList<int> IntCollection = Enumerable.Range(0, 100).ToList().AsReadOnly();
-
-        [Benchmark]
-        public bool ModuloIsEven()
+        var result = false;
+        foreach (var element in IntCollection)
         {
-            var result = false;
-            foreach (var element in IntCollection)
-            {
-                var evenIsEven = element % 2 == 0;
-                var oddIsOdd = element % 2 == 1;
-                result = evenIsEven && oddIsOdd;
-            }
-
-            return result;
+            var evenIsEven = element % 2 == 0;
+            var oddIsOdd = element % 2 == 1;
+            result = evenIsEven && oddIsOdd;
         }
 
-        [Benchmark]
-        public bool BitwiseIsEven()
-        {
-            var result = false; 
-            foreach (var element in IntCollection)
-            {
-                var evenIsEven = (element & 1) == 0;
-                var oddIsOdd = (element & 1) == 1;
-                result = evenIsEven && oddIsOdd;
-            }
+        return result;
+    }
 
-            return result;
+    [Benchmark]
+    public bool BitwiseIsEven()
+    {
+        var result = false;
+        foreach (var element in IntCollection)
+        {
+            var evenIsEven = (element & 1) == 0;
+            var oddIsOdd = (element & 1) == 1;
+            result = evenIsEven && oddIsOdd;
         }
+
+        return result;
     }
 }
