@@ -52,7 +52,7 @@ public class StructBoxTests
     }
 
     [Benchmark]
-    public void ToString()
+    public void AsString()
     {
         var container = new List<string>(_limit);
         for (var i = 0; i < 50; i++)
